nat lib + client added

+++ net/net_message.h
@@ -56,7 +56,7 @@ namespace olc
 				std::memcpy(msg.body.data() + i, &data, sizeof(DataType));
 
 				// Recalculate the message size
-				msg.header.size = msg.size();
+				msg.header.size = (uint32_t) msg.size();
 
 				// Return the target message so it can be chained as with std::cout
 				return msg;
@@ -79,7 +79,7 @@ namespace olc
 				msg.body.resize(i);
 
 				// Recalculate the message size
-				msg.header.size = msg.size();
+				msg.header.size = (uint32_t) msg.size();
 
 				// Return the target message so it can be chained as with std::cout
 				return msg;
